refactor: move resume apis to BaseSyncCall/BaseAsyncCall
